+++ ohseatback/src/main/java/com/ohseat/ohseatback/domain/notice/dto/NoticeDetailResponse.java
@@ -9,11 +9,11 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class NoticeDetailResponse {
 
-    private Long noticeId;
+    private Integer noticeId;
     private String targetBoard;
     private String title;
     private String content;
-    private Long authorId;
+    private Integer authorId;
     private String authorNickName;
     private LocalDateTime createdAt;
     private int views;
